add reader mode Enum class file
