+++ integrations/jina/src/haystack_integrations/components/reader/jina/reader.py
@@ -2,49 +2,14 @@
 # SPDX-FileCopyrightText: 2023-present deepset GmbH <info@deepset.ai>
 #
 # SPDX-License-Identifier: Apache-2.0
+
 import urllib
-from enum import Enum
-from typing import Any, Dict, List, Optional, Union
+from typing import Any, Dict, Union
 
 import requests
 from haystack import Document, component, default_from_dict, default_to_dict
 from haystack.utils import Secret, deserialize_secrets_inplace
-
-
-class JinaReaderMode(Enum):
-    """
-    Enum representing modes for the Jina Reader.
-
-    Modes:
-        READ: For reading documents.
-        SEARCH: For searching within documents.
-        GROUND: For grounding or fact checking.
-
-    """
-
-    READ = "READ"
-    SEARCH = "SEARCH"
-    GROUND = "GROUND"
-
-    def __str__(self):
-        return self.value
-
-    @classmethod
-    def from_str(cls, string: str) -> "JinaReaderMode":
-        """
-        Create the reader mode from a string.
-
-        :param string:
-            String to convert.
-        :returns:
-            Reader mode.
-        """
-        enum_map = {e.value: e for e in JinaReaderMode}
-        reader_mode = enum_map.get(string)
-        if reader_mode is None:
-            msg = f"Unknown reader mode '{string}'. Supported modes are: {list(enum_map.keys())}"
-            raise ValueError(msg)
-        return reader_mode
+from reader_mode import JinaReaderMode
 
 
 @component
@@ -93,13 +58,12 @@ class JinaReader:
 
     @component.output_types(document=Document)
     def run(self, query: str):
-        # check input depending on mode
         mode_map = {"READ": "r", "SEARCH": "s", "GROUND": "g"}
         mode = mode_map[self.mode]
         base_url = f"https://{mode}.jina.ai/"
         encoded_target = urllib.parse.quote(query, safe="")
         url = f"{base_url}{encoded_target}"
         response = self._session.get(url)
-        metadata = {"content_type": response.headers["Content-Type"], "url": input}
+        metadata = {"content_type": response.headers["Content-Type"], "query": query}
         document = [Document(content=response.content, meta=metadata)]
         return document
